Add commands `current_task` and `register`.
These commands communicate with the firebase database attached to this
bot in order to persist the state of the user's task progress. The
structure for the same is a `current_task` field in the database,
corresponding to the user's discord ID, containing an array of objects
which correspond to the user's progress in a specific department.

Everything until now has been designed with very high extensibility in
mind with respect to tasks given by departments. We should be able to
add departments on the fly without changes to the codebase (except for
the tasks.yaml file).