Introduce environment categorisation.
I anticipate that in the future, this will make it much easier to
develop this bot from your local machine with separate environments
in the database and in the code.

This also changes the command that we use for different environments.
This is done so that we don't trigger a command for the `prod`
environment when we're testing the app for `dev`.

Also encapsulate `BOT_TOKEN` as an env variable.